* core/templates/core/post_form.html:         - Added type of the button.
This is needed as the summernote wysiwyg plugin is using a hidden
input field which is required and another div to write the actual
text. On clicking the submit button Chrome tries to validate all
inputs and fails because the input field for the text is empty.

If the button type is not specified the default is "submit". If
changed to type "button" no validation is done.

Signed-off-by: Dimitar Ianakiev <root@mitaka-g.net>